+++ src/backend/base/langflow/base/io/text.py
@@ -3,7 +3,7 @@ from typing import Optional
 from langflow.custom import CustomComponent
 from langflow.field_typing import Text
 from langflow.helpers.record import records_to_text
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class TextComponent(CustomComponent):
