+++ src/backend/base/langflow/base/prompts/utils.py
@@ -1,10 +1,9 @@
-import base64
 from copy import deepcopy
 
 from langchain_core.documents import Document
 
 from langflow.schema import Record
-from langflow.services.deps import get_storage_service
+from langflow.schema.image import get_file_paths
 
 
 def record_to_string(record: Record) -> str:
@@ -48,30 +47,6 @@ async def dict_values_to_string(d: dict) -> dict:
         elif isinstance(value, Document):
             d_copy[key] = document_to_string(value)
     return d_copy
-
-
-async def get_file_paths(files: list[str]):
-    storage_service = get_storage_service()
-    file_paths = []
-    for file in files:
-        flow_id, file_name = file.split("/")
-        file_paths.append(storage_service.build_full_path(flow_id=flow_id, file_name=file_name))
-    return file_paths
-
-
-async def get_files(
-    file_paths: str,
-    convert_to_base64: bool = False,
-):
-    storage_service = get_storage_service()
-    file_objects = []
-    for file_path in file_paths:
-        flow_id, file_name = file_path.split("/")
-        file_object = await storage_service.get_file(flow_id=flow_id, file_name=file_name)
-        if convert_to_base64:
-            file_object = base64.b64encode(file_object).decode("utf-8")
-        file_objects.append(file_object)
-    return file_objects
 
 
 def document_to_string(document: Document) -> str:
