+++ src/backend/base/langflow/components/agents/XMLAgent.py
@@ -3,10 +3,9 @@ from typing import List, Optional
 from langchain.agents import create_xml_agent
 from langchain_core.prompts import ChatPromptTemplate
 
-
 from langflow.base.agents.agent import LCAgentComponent
 from langflow.field_typing import BaseLanguageModel, Text, Tool
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class XMLAgentComponent(LCAgentComponent):
