+++ src/backend/base/langflow/components/textsplitters/LanguageRecursiveTextSplitter.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 from langchain_text_splitters import Language, RecursiveCharacterTextSplitter
 
 from langflow.custom import CustomComponent
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class LanguageRecursiveTextSplitterComponent(CustomComponent):
