+++ src/backend/base/langflow/components/vectorstores/AstraDB.py
@@ -1,11 +1,12 @@
 from typing import List, Optional, Union
+
 from langchain_astradb import AstraDBVectorStore
 from langchain_astradb.utils.astradb import SetupMode
+from langchain_core.retrievers import BaseRetriever
 
 from langflow.custom import CustomComponent
 from langflow.field_typing import Embeddings, VectorStore
 from langflow.schema import Record
-from langchain_core.retrievers import BaseRetriever
 
 
 class AstraDBVectorStoreComponent(CustomComponent):
@@ -155,4 +156,5 @@ class AstraDBVectorStoreComponent(CustomComponent):
                 collection_indexing_policy=collection_indexing_policy,
             )
 
+        return vector_store
         return vector_store
