+++ src/backend/base/langflow/components/vectorstores/MongoDBAtlasVector.py
@@ -4,7 +4,7 @@ from langchain_community.vectorstores.mongodb_atlas import MongoDBAtlasVectorSea
 
 from langflow.custom import CustomComponent
 from langflow.field_typing import Embeddings
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class MongoDBAtlasComponent(CustomComponent):
