+++ src/backend/base/langflow/components/vectorstores/Pinecone.py
@@ -8,7 +8,7 @@ from langchain_pinecone.vectorstores import PineconeVectorStore
 
 from langflow.custom import CustomComponent
 from langflow.field_typing import Embeddings
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class PineconeComponent(CustomComponent):
