+++ src/backend/base/langflow/components/vectorstores/Vectara.py
@@ -9,7 +9,7 @@ from langchain_core.vectorstores import VectorStore
 
 from langflow.custom import CustomComponent
 from langflow.field_typing import BaseRetriever
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class VectaraComponent(CustomComponent):
