+++ src/backend/base/langflow/custom/custom_component/custom_component.py
@@ -7,6 +7,7 @@ import yaml
 from cachetools import TTLCache, cachedmethod
 from langchain_core.documents import Document
 from pydantic import BaseModel
+
 from langflow.custom.code_parser.utils import (
     extract_inner_type_from_generic_alias,
     extract_union_types_from_generic_alias,
