+++ src/backend/base/langflow/graph/utils.py
@@ -2,11 +2,10 @@ from enum import Enum
 from typing import Any, Generator, Union
 
 from langchain_core.documents import Document
-from langflow.schema.schema import Record
 from pydantic import BaseModel
 
 from langflow.interface.utils import extract_input_variables_from_prompt
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 
 class UnbuiltObject:
