+++ src/backend/base/langflow/interface/initialize/loading.py
@@ -8,7 +8,7 @@ from loguru import logger
 
 from langflow.custom.eval import eval_custom_component_code
 from langflow.graph.utils import get_artifact_type, post_process_raw
-from langflow.schema.schema import Record
+from langflow.schema import Record
 
 if TYPE_CHECKING:
     from langflow.custom import CustomComponent
